requirements.txt
# Spotify Voice Control: offline voice to Windows media keys
vosk>=0.3.45
sounddevice>=0.4.6
numpy>=1.24.0
pystray>=0.19.0
Pillow>=10.0.0
spotipy>=2.23.0
